split out transformer into own trait
