+++ pipeline/transformer/transformer_calm/src/main/scala/uk/ac/wellcome/platform/transformer/calm/Transformer.scala
@@ -1,0 +1,8 @@
+package uk.ac.wellcome.platform.transformer.calm
+
+trait Transformer[In, Out] {
+  def transform(in: In): Either[Throwable, Out]
+}
+
+
+
