+++ pipeline/transformer/transformer_calm/src/main/scala/uk/ac/wellcome/platform/transformer/calm/service/CalmTransformerWorkerService.scala
@@ -6,11 +6,8 @@ import uk.ac.wellcome.bigmessaging.BigMessageSender
 import uk.ac.wellcome.json.JsonUtil._
 import uk.ac.wellcome.messaging.sns.{NotificationMessage, SNSConfig}
 import uk.ac.wellcome.messaging.sqs.SQSStream
-import uk.ac.wellcome.models.work.internal.{
-  TransformedBaseWork,
-  UnidentifiedWork,
-  WorkData
-}
+import uk.ac.wellcome.models.work.internal.{TransformedBaseWork, UnidentifiedWork, WorkData}
+import uk.ac.wellcome.platform.transformer.calm.{CalmTransformer, CalmTransformerError, Transformer}
 import uk.ac.wellcome.platform.transformer.calm.models.CalmRecord
 import uk.ac.wellcome.platform.transformer.calm.transformers.CalmToSourceIdentifier
 import uk.ac.wellcome.storage.store.{Readable, VersionedStore}
@@ -18,30 +15,6 @@ import uk.ac.wellcome.storage.{Identified, ObjectLocation, ReadError, Version}
 import uk.ac.wellcome.typesafe.Runnable
 
 import scala.concurrent.Future
-
-trait CalmTransformerError extends Throwable
-case object SourceIdentifierMissingError extends CalmTransformerError
-case object MultipleSourceIdentifiersFoundError extends CalmTransformerError
-
-object CalmTransformer {
-  def transform(calmRecord: CalmRecord)
-    : Either[CalmTransformerError, TransformedBaseWork] = {
-    val sourceIdentifier = CalmToSourceIdentifier.transform(calmRecord)
-
-    sourceIdentifier.map {
-      case head :: Nil =>
-        Right(
-          UnidentifiedWork(
-            version = 1,
-            sourceIdentifier = head,
-            data = WorkData()))
-
-      case _ => Left(MultipleSourceIdentifiersFoundError)
-    } getOrElse Left(SourceIdentifierMissingError)
-
-  }
-
-}
 
 sealed trait CalmTransformerWorkerServiceState {
   val err: Throwable
